+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/profiling/ebpf/EBPFProfilingQueryService.java
@@ -18,49 +18,61 @@
 
 package org.apache.skywalking.oap.server.core.profiling.ebpf;
 
+import com.google.gson.Gson;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.skywalking.oap.server.core.CoreModuleConfig;
+import org.apache.skywalking.oap.server.core.analysis.DownSampling;
 import org.apache.skywalking.oap.server.core.analysis.IDManager;
 import org.apache.skywalking.oap.server.core.analysis.Layer;
+import org.apache.skywalking.oap.server.core.analysis.TimeBucket;
 import org.apache.skywalking.oap.server.core.analysis.manual.process.ProcessDetectType;
 import org.apache.skywalking.oap.server.core.analysis.manual.process.ProcessTraffic;
 import org.apache.skywalking.oap.server.core.analysis.metrics.Metrics;
 import org.apache.skywalking.oap.server.core.profiling.ebpf.analyze.EBPFProfilingAnalyzer;
 import org.apache.skywalking.oap.server.core.query.input.Duration;
-import org.apache.skywalking.oap.server.core.query.input.EBPFProfilingTaskCondition;
 import org.apache.skywalking.oap.server.core.query.type.Attribute;
 import org.apache.skywalking.oap.server.core.query.type.EBPFProfilingAnalyzation;
 import org.apache.skywalking.oap.server.core.query.type.EBPFProfilingAnalyzeTimeRange;
 import org.apache.skywalking.oap.server.core.query.type.EBPFProfilingSchedule;
 import org.apache.skywalking.oap.server.core.query.type.EBPFProfilingTask;
+import org.apache.skywalking.oap.server.core.query.type.EBPFProfilingTaskPrepare;
 import org.apache.skywalking.oap.server.core.query.type.Process;
 import org.apache.skywalking.oap.server.core.storage.IMetricsDAO;
 import org.apache.skywalking.oap.server.core.storage.StorageDAO;
 import org.apache.skywalking.oap.server.core.storage.StorageModule;
 import org.apache.skywalking.oap.server.core.storage.model.Model;
 import org.apache.skywalking.oap.server.core.storage.model.StorageModels;
-import org.apache.skywalking.oap.server.core.storage.profiling.ebpf.EBPFProfilingProcessFinder;
 import org.apache.skywalking.oap.server.core.storage.profiling.ebpf.IEBPFProfilingScheduleDAO;
 import org.apache.skywalking.oap.server.core.storage.profiling.ebpf.IEBPFProfilingTaskDAO;
+import org.apache.skywalking.oap.server.core.storage.profiling.ebpf.IServiceLabelDAO;
+import org.apache.skywalking.oap.server.core.storage.query.IMetadataQueryDAO;
 import org.apache.skywalking.oap.server.library.module.ModuleManager;
 import org.apache.skywalking.oap.server.library.module.Service;
 import org.apache.skywalking.oap.server.library.util.CollectionUtils;
+import org.apache.skywalking.oap.server.library.util.StringUtil;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
 @Slf4j
 @RequiredArgsConstructor
 public class EBPFProfilingQueryService implements Service {
+    private static final Gson GSON = new Gson();
+
     private final ModuleManager moduleManager;
     private final CoreModuleConfig config;
     private final StorageModels storageModels;
 
+    private IMetadataQueryDAO metadataQueryDAO;
+    private IServiceLabelDAO serviceLabelDAO;
     private IEBPFProfilingTaskDAO taskDAO;
     private IEBPFProfilingScheduleDAO scheduleDAO;
     private EBPFProfilingAnalyzer profilingAnalyzer;
@@ -118,13 +130,49 @@ public class EBPFProfilingQueryService implements Service {
         return profilingAnalyzer;
     }
 
-    public List<EBPFProfilingTask> queryEBPFProfilingTasks(EBPFProfilingTaskCondition condition) throws IOException {
-        return getTaskDAO().queryTasks(EBPFProfilingProcessFinder.builder()
-                        .finderType(condition.getFinderType())
-                        .serviceId(condition.getServiceId())
-                        .instanceId(condition.getInstanceId())
-                        .processIdList(Arrays.asList(condition.getProcessId()))
-                .build(), null, 0, 0);
+    public IMetadataQueryDAO getMetadataQueryDAO() {
+        if (metadataQueryDAO == null) {
+            metadataQueryDAO = moduleManager.find(StorageModule.NAME)
+                    .provider()
+                    .getService(IMetadataQueryDAO.class);
+        }
+        return metadataQueryDAO;
+    }
+
+    public IServiceLabelDAO getServiceLabelDAO() {
+        if (serviceLabelDAO == null) {
+            serviceLabelDAO = moduleManager.find(StorageModule.NAME)
+                    .provider()
+                    .getService(IServiceLabelDAO.class);
+        }
+        return serviceLabelDAO;
+    }
+
+    public EBPFProfilingTaskPrepare queryPrepareCreateEBPFProfilingTaskData(String serviceId) throws IOException {
+        final EBPFProfilingTaskPrepare prepare = new EBPFProfilingTaskPrepare();
+        // query process count in last 10 minutes
+        final long endTimestamp = System.currentTimeMillis();
+        final long startTimestamp = endTimestamp - TimeUnit.MINUTES.toMillis(10);
+        final long processesCount = getMetadataQueryDAO().getProcessesCount(serviceId, null, null,
+                TimeBucket.getTimeBucket(startTimestamp, DownSampling.Minute),
+                TimeBucket.getTimeBucket(endTimestamp, DownSampling.Minute));
+        if (processesCount <= 0) {
+            prepare.setCouldProfiling(false);
+            prepare.setProcessLabels(Collections.emptyList());
+            return prepare;
+        }
+        prepare.setCouldProfiling(true);
+        final List<String> processLabels = getServiceLabelDAO().queryAllLabels(serviceId);
+        if (processLabels != null && !processLabels.isEmpty()) {
+            prepare.setProcessLabels(processLabels.stream().distinct().collect(Collectors.toList()));
+        } else {
+            prepare.setProcessLabels(Collections.emptyList());
+        }
+        return prepare;
+    }
+
+    public List<EBPFProfilingTask> queryEBPFProfilingTasks(String serviceId) throws IOException {
+        return getTaskDAO().queryTasks(Arrays.asList(serviceId), null, 0, 0);
     }
 
     public List<EBPFProfilingSchedule> queryEBPFProfilingSchedules(String taskId, Duration duration) throws IOException {
@@ -147,8 +195,8 @@ public class EBPFProfilingQueryService implements Service {
         return schedules;
     }
 
-    public EBPFProfilingAnalyzation getEBPFProfilingAnalyzation(String taskId, List<EBPFProfilingAnalyzeTimeRange> timeRanges) throws IOException {
-        return getProfilingAnalyzer().analyze(taskId, timeRanges);
+    public EBPFProfilingAnalyzation getEBPFProfilingAnalyzation(List<String> scheduleIdList, List<EBPFProfilingAnalyzeTimeRange> timeRanges) throws IOException {
+        return getProfilingAnalyzer().analyze(scheduleIdList, timeRanges);
     }
 
     private Process convertProcess(ProcessTraffic traffic) {
@@ -168,6 +216,9 @@ public class EBPFProfilingQueryService implements Service {
             for (String key : traffic.getProperties().keySet()) {
                 process.getAttributes().add(new Attribute(key, traffic.getProperties().get(key).getAsString()));
             }
+        }
+        if (StringUtil.isNotEmpty(traffic.getLabelsJson())) {
+            process.getLabels().addAll(GSON.<List<String>>fromJson(traffic.getLabelsJson(), ArrayList.class));
         }
         return process;
     }