+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/profiling/ebpf/storage/EBPFProfilingTaskRecord.java
@@ -24,7 +24,6 @@ import lombok.Data;
 import org.apache.skywalking.oap.server.core.analysis.Stream;
 import org.apache.skywalking.oap.server.core.analysis.config.NoneStream;
 import org.apache.skywalking.oap.server.core.analysis.worker.NoneStreamProcessor;
-import org.apache.skywalking.oap.server.core.query.type.EBPFProfilingProcessFinderType;
 import org.apache.skywalking.oap.server.core.source.ScopeDeclaration;
 import org.apache.skywalking.oap.server.core.storage.annotation.Column;
 import org.apache.skywalking.oap.server.core.storage.type.Convert2Entity;
@@ -41,13 +40,9 @@ import static org.apache.skywalking.oap.server.core.source.DefaultScopeDefine.EB
 @Stream(name = EBPFProfilingTaskRecord.INDEX_NAME, scopeId = EBPF_PROFILING_TASK,
         builder = EBPFProfilingTaskRecord.Builder.class, processor = NoneStreamProcessor.class)
 public class EBPFProfilingTaskRecord extends NoneStream {
-
     public static final String INDEX_NAME = "ebpf_profiling_task";
-    public static final String PROCESS_FIND_TYPE = "process_find_type";
     public static final String SERVICE_ID = "service_id";
-    public static final String INSTANCE_ID = "instance_id";
-    public static final String PROCESS_ID = "process_id";
-    public static final String PROCESS_NAME = "process_name";
+    public static final String PROCESS_LABELS_JSON = "process_labels_json";
     public static final String START_TIME = "start_time";
     public static final String TRIGGER_TYPE = "trigger_type";
     public static final String FIXED_TRIGGER_DURATION = "fixed_trigger_duration";
@@ -55,16 +50,12 @@ public class EBPFProfilingTaskRecord extends NoneStream {
     public static final String CREATE_TIME = "create_time";
     public static final String LAST_UPDATE_TIME = "last_update_time";
 
-    @Column(columnName = PROCESS_FIND_TYPE)
-    private int processFindType = EBPFProfilingProcessFinderType.UNKNOWN.value();
+    public static final int PROCESS_LABELS_JSON_MAX_LENGTH = 1000;
+
     @Column(columnName = SERVICE_ID)
     private String serviceId;
-    @Column(columnName = INSTANCE_ID, length = 600)
-    private String instanceId;
-    @Column(columnName = PROCESS_ID, length = 600)
-    private String processId;
-    @Column(columnName = PROCESS_NAME, length = 500)
-    private String processName;
+    @Column(columnName = PROCESS_LABELS_JSON, length = PROCESS_LABELS_JSON_MAX_LENGTH)
+    private String processLabelsJson;
     @Column(columnName = START_TIME)
     private long startTime;
     @Column(columnName = TRIGGER_TYPE)
@@ -81,9 +72,9 @@ public class EBPFProfilingTaskRecord extends NoneStream {
     @Override
     public String id() {
         return Hashing.sha256().newHasher()
-                .putString(processId, Charsets.UTF_8)
+                .putString(serviceId, Charsets.UTF_8)
+                .putString(processLabelsJson, Charsets.UTF_8)
                 .putLong(createTime)
-                .putInt(processFindType)
                 .hash().toString();
     }
 
@@ -92,11 +83,8 @@ public class EBPFProfilingTaskRecord extends NoneStream {
         @Override
         public EBPFProfilingTaskRecord storage2Entity(final Convert2Entity converter) {
             final EBPFProfilingTaskRecord record = new EBPFProfilingTaskRecord();
-            record.setProcessFindType(((Number) converter.get(PROCESS_FIND_TYPE)).intValue());
             record.setServiceId((String) converter.get(SERVICE_ID));
-            record.setInstanceId((String) converter.get(INSTANCE_ID));
-            record.setProcessId((String) converter.get(PROCESS_ID));
-            record.setProcessName((String) converter.get(PROCESS_NAME));
+            record.setProcessLabelsJson((String) converter.get(PROCESS_LABELS_JSON));
             record.setTriggerType(((Number) converter.get(TRIGGER_TYPE)).intValue());
             record.setStartTime(((Number) converter.get(START_TIME)).longValue());
             record.setFixedTriggerDuration(((Number) converter.get(FIXED_TRIGGER_DURATION)).longValue());
@@ -109,11 +97,8 @@ public class EBPFProfilingTaskRecord extends NoneStream {
 
         @Override
         public void entity2Storage(final EBPFProfilingTaskRecord storageData, final Convert2Storage converter) {
-            converter.accept(PROCESS_FIND_TYPE, storageData.getProcessFindType());
             converter.accept(SERVICE_ID, storageData.getServiceId());
-            converter.accept(INSTANCE_ID, storageData.getInstanceId());
-            converter.accept(PROCESS_ID, storageData.getProcessId());
-            converter.accept(PROCESS_NAME, storageData.getProcessName());
+            converter.accept(PROCESS_LABELS_JSON, storageData.getProcessLabelsJson());
             converter.accept(TRIGGER_TYPE, storageData.getTriggerType());
             converter.accept(START_TIME, storageData.getStartTime());
             converter.accept(FIXED_TRIGGER_DURATION, storageData.getFixedTriggerDuration());