+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/storage/StorageModule.java
@@ -20,6 +20,7 @@ package org.apache.skywalking.oap.server.core.storage;
 
 import org.apache.skywalking.oap.server.core.storage.cache.INetworkAddressAliasDAO;
 import org.apache.skywalking.oap.server.core.storage.management.UITemplateManagementDAO;
+import org.apache.skywalking.oap.server.core.storage.profiling.ebpf.IServiceLabelDAO;
 import org.apache.skywalking.oap.server.core.storage.profiling.trace.IProfileTaskLogQueryDAO;
 import org.apache.skywalking.oap.server.core.storage.profiling.trace.IProfileTaskQueryDAO;
 import org.apache.skywalking.oap.server.core.storage.profiling.trace.IProfileThreadSnapshotQueryDAO;
@@ -74,7 +75,8 @@ public class StorageModule extends ModuleDefine {
             IEventQueryDAO.class,
             IEBPFProfilingTaskDAO.class,
             IEBPFProfilingScheduleDAO.class,
-            IEBPFProfilingDataDAO.class
+            IEBPFProfilingDataDAO.class,
+            IServiceLabelDAO.class
         };
     }
 }