+++ oap-server/server-storage-plugin/storage-iotdb-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/iotdb/query/IoTDBEBPFProfilingTaskDAO.java
@@ -18,25 +18,24 @@
 
 package org.apache.skywalking.oap.server.storage.plugin.iotdb.query;
 
+import com.google.gson.Gson;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.skywalking.oap.server.core.analysis.IDManager;
-import org.apache.skywalking.oap.server.core.query.type.EBPFProfilingProcessFinderType;
 import org.apache.skywalking.oap.server.core.profiling.ebpf.storage.EBPFProfilingTargetType;
 import org.apache.skywalking.oap.server.core.profiling.ebpf.storage.EBPFProfilingTaskRecord;
 import org.apache.skywalking.oap.server.core.profiling.ebpf.storage.EBPFProfilingTriggerType;
 import org.apache.skywalking.oap.server.core.query.type.EBPFProfilingTask;
 import org.apache.skywalking.oap.server.core.storage.StorageData;
-import org.apache.skywalking.oap.server.core.storage.profiling.ebpf.EBPFProfilingProcessFinder;
 import org.apache.skywalking.oap.server.core.storage.profiling.ebpf.IEBPFProfilingTaskDAO;
 import org.apache.skywalking.oap.server.core.storage.type.StorageBuilder;
-import org.apache.skywalking.oap.server.library.util.CollectionUtils;
 import org.apache.skywalking.oap.server.library.util.StringUtil;
 import org.apache.skywalking.oap.server.storage.plugin.iotdb.IoTDBClient;
 import org.apache.skywalking.oap.server.storage.plugin.iotdb.IoTDBIndexes;
 
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -45,35 +44,25 @@ import org.apache.skywalking.oap.server.storage.plugin.iotdb.utils.IoTDBUtils;
 @Slf4j
 @RequiredArgsConstructor
 public class IoTDBEBPFProfilingTaskDAO implements IEBPFProfilingTaskDAO {
+    private static final Gson GSON = new Gson();
+
     private final IoTDBClient client;
     private final StorageBuilder<EBPFProfilingTaskRecord> storageBuilder = new EBPFProfilingTaskRecord.Builder();
 
     @Override
-    public List<EBPFProfilingTask> queryTasks(EBPFProfilingProcessFinder finder,
+    public List<EBPFProfilingTask> queryTasks(List<String> serviceIdList,
                                               EBPFProfilingTargetType targetType,
                                               long taskStartTime, long latestUpdateTime)
             throws IOException {
         StringBuilder query = new StringBuilder();
         query.append("select * from ");
         IoTDBUtils.addModelPath(client.getStorageGroup(), query, EBPFProfilingTaskRecord.INDEX_NAME);
-        Map<String, String> indexAndValueMap = new HashMap<>();
-        if (StringUtil.isNotEmpty(finder.getServiceId())) {
-            indexAndValueMap.put(IoTDBIndexes.SERVICE_ID_IDX, finder.getServiceId());
-        }
-        if (StringUtil.isNotEmpty(finder.getInstanceId())) {
-            indexAndValueMap.put(IoTDBIndexes.INSTANCE_ID_INX, finder.getInstanceId());
-        }
-        if (CollectionUtils.isNotEmpty(finder.getProcessIdList())) {
-            final List<String> processIdList = finder.getProcessIdList();
-            for (int i = 0; i < processIdList.size(); i++) {
-                if (i > 0) {
-                    query.append(", ");
-                }
-                final HashMap<String, String> indexWithProcessId = new HashMap<>(indexAndValueMap);
-                indexWithProcessId.put(IoTDBIndexes.PROCESS_ID_INX, processIdList.get(i));
-                IoTDBUtils.addQueryIndexValue(EBPFProfilingTaskRecord.INDEX_NAME, query, indexWithProcessId);
+        for (int i = 0; i < serviceIdList.size(); i++) {
+            if (i > 0) {
+                query.append(", ");
             }
-        } else {
+            Map<String, String> indexAndValueMap = new HashMap<>();
+            indexAndValueMap.put(IoTDBIndexes.SERVICE_ID_IDX, serviceIdList.get(i));
             IoTDBUtils.addQueryIndexValue(EBPFProfilingTaskRecord.INDEX_NAME, query, indexAndValueMap);
         }
 
@@ -105,13 +94,14 @@ public class IoTDBEBPFProfilingTaskDAO implements IEBPFProfilingTaskDAO {
     private EBPFProfilingTask parseTask(EBPFProfilingTaskRecord record) {
         final EBPFProfilingTask task = new EBPFProfilingTask();
         task.setTaskId(record.id());
-        task.setProcessFinderType(EBPFProfilingProcessFinderType.valueOf(record.getProcessFindType()));
         task.setServiceId(record.getServiceId());
         task.setServiceName(IDManager.ServiceID.analysisId(record.getServiceId()).getName());
-        task.setInstanceId(record.getInstanceId());
-        task.setInstanceName(IDManager.ServiceInstanceID.analysisId(record.getInstanceId()).getName());
-        task.setProcessId(record.getProcessId());
-        task.setProcessName(record.getProcessName());
+        final String processLabelsJson = record.getProcessLabelsJson();
+        if (StringUtil.isNotEmpty(processLabelsJson)) {
+            task.setProcessLabels(GSON.<List<String>>fromJson(processLabelsJson, ArrayList.class));
+        } else {
+            task.setProcessLabels(Collections.emptyList());
+        }
         task.setTaskStartTime(record.getStartTime());
         task.setTriggerType(EBPFProfilingTriggerType.valueOf(record.getTriggerType()));
         task.setFixedTriggerDuration(record.getFixedTriggerDuration());