+++ docs/en/changes/changes.md
@@ -13,6 +13,8 @@
 * Support BanyanDB global index for entities. Log and Segment record entities declare this new feature.
 * Remove unnecessary analyzer settings in columns of templates. Many were added due to analyzer's default value.
 * Simplify the Kafka Fetch configuration in cluster mode.
+* [Breaking Change] Update the eBPF Profiling task to the service level,
+  please delete index/table: `ebpf_profiling_task`, `process_traffic`.
 
 #### UI